Update list manager to work for local dev outside of WP (#674)
* update list manager to work outside of WP

+++ wordpress/wp-content/plugins/cds-base/classes/Modules/ListManager/app/src/components/DeleteActionLink.tsx
@@ -5,16 +5,24 @@ import { useList } from "../store/ListContext";
 import { ConfirmActionLink } from "./ConfirmActionLink"
 
 
-export const DeleteActionLink = ({id = ''}:{id: string}) => {
+export const DeleteActionLink = ({ id = '' }: { id: string }) => {
     const { request, response } = useFetch({ data: [] })
     const { dispatch } = useList();
 
     const deleteList = async ({ id = '' }: { id: string }) => {
-        await request.delete(`/list/${id}`)
 
-        if (response.ok) {
+        if (process.env.NODE_ENV !== "development") {
+            await request.delete(`/list/${id}`);
+
+            if (response.ok) {
+                dispatch({ type: "delete", payload: { id } });
+            }
+        } else {
+            // for local dev --- dispatch the delete
             dispatch({ type: "delete", payload: { id } });
         }
+
+
     }
 
     return <ConfirmActionLink text={"Delete"} isConfirmedHandler={() => deleteList({ id })} />
